+++ src/main/java/org/gbif/common/messaging/api/messages/VerbatimPersistedMessage.java
@@ -25,14 +25,14 @@ public class VerbatimPersistedMessage implements DatasetBasedMessage {
 
   private final OccurrencePersistenceStatus status;
 
-  private final int occurrenceKey;
+  private final long occurrenceKey;
 
   @JsonCreator
   public VerbatimPersistedMessage(
     @JsonProperty("datasetUuid") UUID datasetUuid,
     @JsonProperty("attempt") int attempt,
     @JsonProperty("status") OccurrencePersistenceStatus status,
-    @JsonProperty("occurrenceKey") int occurrenceKey
+    @JsonProperty("occurrenceKey") long occurrenceKey
   ) {
     this.datasetUuid = checkNotNull(datasetUuid, "datasetUuid can't be null");
     checkArgument(attempt > 0, "attempt must be greater than 0");
@@ -59,7 +59,7 @@ public class VerbatimPersistedMessage implements DatasetBasedMessage {
     return status;
   }
 
-  public int getOccurrenceKey() {
+  public long getOccurrenceKey() {
     return occurrenceKey;
   }
